pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anthrokit"
version = "1.0.0"
description = "Token-based anthropomorphism design system for conversational AI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AnthroKit Contributors", email = "anthrokit@example.com"}
]
maintainers = [
    {name = "AnthroKit Contributors", email = "anthrokit@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "anthropomorphism",
    "conversational-ai",
    "chatbot",
    "design-system",
    "explainable-ai",
    "responsible-ai",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# API service dependencies
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
]

# LLM stylization dependencies
llm = [
    "openai>=1.0.0",
]

# Development dependencies
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.990",
    "isort>=5.10",
]

# XAIagent experiment dependencies
experiment = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "shap>=0.42.0",
]

# Complete installation (all features)
all = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
    "openai>=1.0.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.990",
    "isort>=5.10",
]

[project.urls]
Homepage = "https://github.com/your-org/anthrokit"
Documentation = "https://anthrokit.readthedocs.io"
Repository = "https://github.com/your-org/anthrokit"
"Bug Tracker" = "https://github.com/your-org/anthrokit/issues"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["anthrokit", "anthrokit.api"]

[tool.setuptools.package-data]
anthrokit = ["anthrokit.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=anthrokit --cov-report=term-missing"
testpaths = ["experiment/tests"]
pythonpath = ["."]
